Move new test to new integration_tests.go file
